Extract `FeedImageDataLoader` in its own file
